10824.cpp

+++ boj/string/10824.cpp
@@ -15,5 +15,5 @@ int main(void){
 		cin >> tmp;
 		cd += tmp;
 	}
-	cout << stoi(ab) + stoi(cd);
+	cout << stoll(ab) + stoll(cd);
 }
